Arbitrary custom fields on a subscriber
The UI requires some JS implemented to put a new template on the page,
but the basic concept behind arbitrary custom fields on a subscriber is
in place now.

Likely going to need some new logic to handle extracting the data from
the custom fields into the templates, as well as what to do when that
data isn’t present (the fields are arbitrary after all, so not all
subscribers are going to have them).